% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KRRC_VWG}
\alias{KRRC_VWG}
\title{Kernel Ridge Regreggion Classifier with Veronese Whitney Gaussian}
\usage{
KRRC_VWG(x, xtest, y, Sig, lambda)
}
\arguments{
\item{x}{Training Array for building subspaces}

\item{xtest}{Test set (Array)}

\item{y}{Test label}

\item{Sig}{the regularization parameter for kernel}

\item{lambda}{the regularization parameter for Ridge Reg}
}
\value{
Est_Mat distance matrix
}
\description{
Kernel Ridge Regreggion Classifier with Veronese Whitney Gaussian
}
