% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRC.R
\name{RRC1}
\alias{RRC1}
\title{Naive Ridge Regression Classifier}
\usage{
RRC1(sub_X, Test_X, Test_Y, lambda)
}
\arguments{
\item{sub_X}{Training data for building subspaces}

\item{Test_X}{Test data}

\item{Test_Y}{Label of Test set}

\item{labmda}{Regularization parameter for Ridge Reg}
}
\value{
Est_RRC_label Estimated label
}
\description{
Naive Ridge Regression Classifier
}
