\name{ShapeKRRC-package}
\alias{ShapeKRRC-package}
\alias{ShapeKRRC}
\docType{package}
\title{
\packageTitle{ShapeKRRC}
}
\description{
\packageDescription{ShapeKRRC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShapeKRRC}
\packageIndices{ShapeKRRC}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{ShapeKRRC}

Maintainer: \packageMaintainer{ShapeKRRC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
